Crate cc_consensus_poa

Proof of Access consensus

Introduction

Proof of Access is a kind of lightweight storage consensus initially adopted by Arweave. In arweave, PoA serves as an enhancement of Proof of Work in which the entire recall block data is included in the material to be hashed for input to the proof of work.

Requiring ProofOfAccess incentivises storage as miners need access to random blocks from the blockweave’s history in order to mine new blocks and receive mining rewards.

Overview

The general workflow of PoA is described briefly below:

Pick a random byte from the whole network storage (BlockWeave).

The block weave can be seen as an ever growing gigantic array.

Currently, the randome byte is determined by hashing the parent header hash for N times(see calculate_challenge_byte), which will be replaced with another hashing strategy in SPoRA.

Locate the extrinsic in which the random byte is included.

Check if the data of extrinsic located in Step 2 exists in the local storage.

If the data does exist locally, create the two merkle proofs of extrinsic and data chunks respectively.

If not, repeat from Step 1 by choosing another random byte.

Usage

Normally, PoA needs to be used with other consensus algorithem like PoW or PoS together as it’s not typically designed for solving the problem of selecting one from the validator set to author next block in an unpredictable or fair way. In another word, PoA is usually exploited as a precondition for PoW or PoS in order to encourage the miners to store more data locally.

This crate implements the core algorithem of Proof of Access in construct_poa and provides the inherent data provider via PoaInherentDataProvider. PurePoaBlockImport implements the BlockImport trait, thus can be wrapped in another block importer.

To use this engine, you can create an inhehrent extrinsic using the data provided by PoaInherentDataProvider in a pallet. Furthermore, you need to wrap the PurePoaBlockImport into your existing block import pipeline. Refer to the Substrate docs for more information about creating a nested BlockImport.

Structs

ChunkProof

This struct includes the raw bytes of recall chunk as well as the chunk proof stuffs.

ChunkProofBuilder

A builder for creating a ChunkProof from the raw transaction data.

ChunkProofVerifier

A verifier for chunk proof.

PoaBuilder

A builder for creating PoaOutcome.

PoaConfiguration

Configuration of the PoA consensus engine.

PoaInherentDataProvider

A type for creating the inherent data for pallet poa.

ProofOfAccess

This struct is used to prove the random historical data access of block author.

PurePoaBlockImport

A pure block importer for PoA.

RecallInfo

All information of recall block that is required to build a ProofOfAccess.

TxProofVerifier

A verifier for tx proof.

Enums

Error

Error type for poa consensus.

PoaOutcome

This struct represents the outcome of creating the inherent data of ProofOfAccess.

Constants

POA_ENGINE_ID

The engine id for the Proof of Access consensus.

Functions

build_extrinsic_proof

Returns the calculated merkle proof given extrinsic_index and extrinsics_root.

calculate_challenge_byte

Returns the position of recall byte in the entire weave.

construct_poa

Returns a PoaOutcome after the poa construction.

verify_chunk_proof

Verifies the chunk matches given the chunk_root and proof.

verify_extrinsic_proof

Verify the extrinsic proof on the extrinsics root and related encoded extrinsic.