Crate cc_consensus_poa

This crate creates the inherent data based on the Proof of Access consensus.

TODO: verify PoA stored in the block header.

Structs

ChunkProof
This struct includes the raw bytes of recall chunk as well as the chunk proof stuffs.

ChunkProofBuilder
Utility for buliding a chunk proof from the raw transaction data.

ChunkProofVerifier
Helper struct to verify the chunk proof.

InherentDataProvider

ProofOfAccess
This struct is used to prove the random historical data access of block author.

Enums

Error

Constants

MAX_CHUNK_PATH
Maximum byte size of chunk merkle path.

MAX_DEPTH
The maximum depth of attempting to generate a valid PoA.

MAX_TX_PATH
Maximum byte size of transaction merkle path.

Functions

build_extrinsic_proof

construct_poa
Constructs a valid Proof of Access.

verify_chunk_proof
Verifies the chunk given the chunk_root and proof.

verify_extrinsic_proof